==> 01.install-dependencies.ipynb <==
# Install Dependencies

Sets up and install python dependencies for the tutorial.

## Modules needed for ELMo.

%pip install update spacy tensorflow tensorflow_hub sklearn xgboost

## Get corpus for Spacy

Get word corpus for spacy.

%run -m spacy download en

## NLTK
Spacy seems to have some issues on windows installing en.
Install NLTK for if running this on windows.

%pip install nltk

import nltk
nltk.download("wordnet")

## Dependencies for Bert

%pip install bert-tensorflow sklearn tensorflow_hub
