==> 05.bert.ipynb <==
# Bert Classification

Uses Bert to classify movie reviews. This example adds classification as a layer to the pre-trained Bert model.

from datetime import datetime
import os

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
import pandas as pd
import tensorflow as tf

## Configuration

data_dir = "data"

tf.logging.set_verbosity(tf.logging.ERROR)

## Load Data

imdb_data = pd.read_pickle("{}/imdb_data.pickle.gz".format(data_dir))

sample_size = 100
seed = 11989

train = imdb_data[imdb_data.data_set.str.lower() == 'train'].sample(sample_size, random_state = seed)
test = imdb_data[imdb_data.data_set.str.lower() == 'test'].sample(sample_size, random_state = seed*2)

print("Training size: {}".format(len(train)))
print("Testing size: {}".format(len(test)))

## Setup Bert

from bert_classifier import *

bc = BertClassifier(data_column='sentence', label_column='polarity')

bc.config

## Training

start = datetime.now()
bc.train(train)
et = datetime.now() - start
print("Training time: {}".format(et))

## Testing

start = datetime.now()
results = bc.test(test)
et = datetime.now() - start
print("Test time: {}".format(et))
print("Results:")
print(pd.DataFrame([results]).iloc[0])

## Get Predictions

start = datetime.now()
predictions = bc.predict(test)
et = datetime.now() - start
print("Prediction time: {}".format(et))

pd.DataFrame(predictions)[0:10]
